requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
gunicorn==22.0.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
networkx==3.3
python-louvain==0.16
Pillow==10.3.0
faker==24.0.0
apscheduler==3.10.4
sentence-transformers==3.0.1
